Fix return type of floor function to int and update function signature

+++ python_variable_annotations/2-floor.py
@@ -4,6 +4,6 @@ Module providing a function that take a float and return the floor of the float
 """
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     """Function that return the floor of the float"""
-    return(n)
+    return int(n)
